Initial Commit
TO DATE:
The Voice Command System handles the user to system interaction by reading voice input and detecting keywords.  Any keywords detected are passed to the Voice Controller.

The Voice Controller handles communication between the VCS and the AI Player Controller.  Any commands that are detected by the VPS are passed into the Voice Controller where they are processed into an appropriate event that the Player Controller will react to.

The Player Controller handles the events that are processed by the Voice Controller and determines the best action to perform in response.  It uses a state machine setup to appropriately handle how the commands are processed depending on its current state.