Updated Character Animations + Animation Events
Updated Archer and Rogue ability designs and animations in line with the previous commit.
Rewrote the way abilities are handled to accommodate the new animation event system, allowing for attacks to be made at a specific point during the animation rather than at the start or end of the animation.
Rogue, Priest, Mage animation events implemented.