New Level Model + UI Display for Room Data + Rendering Optimization
Created a custom mesh for the level more inline with the final design with separated rooms and custom environmental elements to distinguish each room.

Created a new UI display to show to player on screen all the enemies and interactable objects found in the room they're currently in.

Implemented a system that stops rendering a room including all the objects and entities within that room unless a player character is present in the room.